LLM-TradeBot-AngelOne/requirements.txt
# AngelOne SmartAPI
smartapi-python>=1.3.0
pyotp>=2.8.0
pytz>=2023.3

# Data Processing
pandas>=2.1.4
numpy>=1.26.2
ta==0.11.0

# Async Processing
aiohttp==3.9.1
websockets==12.0

# LLM APIs (OpenAI, DeepSeek, Claude, etc.)
openai==1.6.1

# Data Storage
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Configuration Management
pyyaml==6.0.1
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Backtesting and Visualization
backtrader==1.9.78.123
matplotlib==3.8.2
seaborn==0.13.0

# Machine Learning
lightgbm>=4.0.0
scikit-learn>=1.3.0

# Utilities
requests==2.31.0
jsonschema==4.20.0

# Web Dashboard
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Testing
pytest>=7.4.0
hypothesis>=6.0.0
